Fix inappropriate characterData variable name
